add Excepitions

+++ src/main/java/com/example/meu_primeiro_springboot/service/ProdutoService.java
@@ -1,8 +1,8 @@
 package com.example.meu_primeiro_springboot.service;
 
 import java.util.List;
-import java.util.Optional;
 import org.springframework.stereotype.Service;
+import com.example.meu_primeiro_springboot.exceptions.RecursoNaoEncontradoException;
 import com.example.meu_primeiro_springboot.model.Produto;
 import com.example.meu_primeiro_springboot.repository.ProdutoRepository;
 
@@ -19,8 +19,9 @@ public class ProdutoService {
         return produtoRepository.findAll();
     }
 
-    public Optional<Produto> buscarPorId(long id) {
-        return produtoRepository.findById(id);
+    public Produto buscarPorId(long id) {
+        return produtoRepository.findById(id)
+        .orElseThrow(() -> new RecursoNaoEncontradoException("Produto não encontrado com id: " + id));
     }
 
     public Produto salvarProduto(Produto produto) {
@@ -28,6 +29,9 @@ public class ProdutoService {
     }
 
     public void deletarProduto(long id) {
+        if (!produtoRepository.existsById(id)) {
+            throw new RecursoNaoEncontradoException("Produto não encontrado com id: " + id);
+        }
         produtoRepository.deleteById(id);
     }
